pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solanakit"
version = "1.0.1"
description = "Python toolkit for blockchain automation — async requester, Web3 helpers, multi-process runner."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Brandy Good", email = "genoshide@users.noreply.github.com" }]
keywords = ["blockchain", "web3", "automation", "testnet", "bot", "defi", "aiohttp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "colorama>=0.4.6",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "watchdog>=4.0",
    "requests>=2.31",
    "cryptography>=43.0",
]

[project.optional-dependencies]
web3 = [
    "web3>=6.0.0",
    "eth-account>=0.10.0",
    "eth-utils>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/genoshide"
Repository = "https://github.com/genoshide/chainutils"

[tool.hatch.build.targets.wheel]
packages = ["src/solanakit"]

[tool.hatch.metadata]
allow-direct-references = true
